+++ ChangeLog.md
@@ -21,17 +21,21 @@
     and `baker update-stake` so that in Protocol version < 4, they generate the former P3
     transaction, and in Protocol version 4, they generate the relevant "configure baker" transaction.
   - Support raw queries `GetPoolStatus` and `GetBakerList`.
+- Add `consensus show-chain-parameters` subcommand to print the chain parameters.
 
 ## 3.0.4
+
 - Rename `--encrypted` and `--decrypt-encrypted` flags to `account show` to
   `--shielded` and `--reveal-shielded`.
 
 ## 3.0.3
+
 - support the same range of tag names for LEI as is supported by identity
   issuance command line tools. Previously only UNNAMED#13 was supported and now
   both lei and the former are.
 
 ## 3.0.2
+
 - credentials revealing newly introduced attribute LEI can be deployed
 
 ## 3.0.1
@@ -65,15 +69,16 @@
 
 - show smart contract state as raw bytes, if schema is provided but doesn't include the state type.
 - warn about sending transfers with oversized memos.
-- show amount in GTU instead of µGTU when trying to send an encrypted amount that is larger than 
+- show amount in GTU instead of µGTU when trying to send an encrypted amount that is larger than
   the encrypted balance.
 
 ## 1.1.0
+
 - The `account show` command can receive a credential registration ID instead of a name or address.
 - support sending the three new transaction types, i.e. TransferWithMemo, EncryptedTransferWithMemo
   and TransferWithScheduleAndMemo
 - show transfer memo in transaction status
-- show protocolVersion, genesisIndex, currentEraGenesisBlock and currentEraGenesisTime in 
+- show protocolVersion, genesisIndex, currentEraGenesisBlock and currentEraGenesisTime in
   consensus status
 - this version is only compatible with node version 1.1.0 and later.
 - warn the user when trying to deploy a wasm module of size > 65536 bytes
@@ -103,6 +108,7 @@
 - Fix incorrect parsing of `ReceiveName` when the contract field contains a dot.
 
 ## 0.7.0
+
 - Add ContractName and ReceiveName to schema.
 - Add UInt128 and Int128 to schema type.
 - Add 'transaction register-data' command.